refactor file structure of modeling module
