+++ auto_gptq/__init__.py
@@ -1,2 +1,2 @@
 from .modeling import BaseQuantizeConfig
-from .modeling_auto import AutoGPTQForCausalLM
+from .modeling import AutoGPTQForCausalLM
