+++ auto_gptq/modeling/__init__.py
@@ -1,14 +1,7 @@
 from ._base import BaseGPTQForCausalLM, BaseQuantizeConfig
+from .auto import *
 from .bloom import *
 from .gpt_neox import *
 from .gptj import *
 from .llama import *
 from .opt import *
-
-GPTQ_CAUSAL_LM_MODEL_MAP = {
-    "bloom": BloomGPTQForCausalLM,
-    "gpt_neox": GPTNeoXGPTQForCausalLM,
-    "gptj": GPTJGPTQForCausalLM,
-    "llama": LlamaGPTQForCausalLM,
-    "opt": OPTGPTQForCausalLM
-}
