+++ auto_gptq/modeling/_utils.py
@@ -1,7 +1,9 @@
 from logging import getLogger
 
 import torch.nn as nn
+from transformers import AutoConfig
 
+from ._const import SUPPORTED_MODELS
 from ..quantization import make_quant, QuantLinear
 
 logger = getLogger(__name__)
@@ -36,4 +38,12 @@ def pack_model(model, quantizers, bits, group_size):
     logger.info('Model packed.')
 
 
-__all__ = ["find_layers", "get_module_by_name", "pack_model"]
+def check_and_get_model_type(model_dir):
+    config = AutoConfig.from_pretrained(model_dir)
+    if config.model_type not in SUPPORTED_MODELS:
+        raise TypeError(f"{config.model_type} isn't supported yet.")
+    model_type = config.model_type
+    return model_type
+
+
+__all__ = ["find_layers", "get_module_by_name", "pack_model", "check_and_get_model_type"]
